backend/requirements.txt
Flask==3.0.0
tensorflow>=2.15.0
numpy>=1.24.3
Pillow>=10.1.0
opencv-python>=4.8.1.78
SpeechRecognition>=3.10.0
gTTS>=2.5.0
Werkzeug>=3.0.1
PyMySQL>=1.1.0
cryptography>=41.0.7
pydub>=0.25.1
python-dotenv==1.0.0
librosa>=0.10.1
matplotlib>=3.8.2
mysql-connector-python>=8.2.0
requests>=2.31.0
